Update getting-started-with-group-managed-service-accounts.md
The current document has incorrect syntax for many of the commands.  changed <string> to &lt;string&gt; in all scenarios to ensure that the code snippets were displayed correctly.  Added a note about the syntax of the command, as example 1 had some optional switches outside of [square brackets], causing confusion.